extract middleware into its own package
